GenAI/requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# tools
langchain-community
ddgs
langchain_experimental      

scikit-learn

# Anthropic Integration
# langchain-anthropic

# Google Gemini (PaLM) Integration
# langchain-google-genai
# google-generativeai

# Hugging Face Integration
# langchain-huggingface
# transformers
# huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
# numpy